+++ src/irmin/read_only.mli
@@ -14,32 +14,4 @@
  * OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
  *)
 
-open Import
-open Store_properties
-
-module type S = sig
-  (** {1 Append-only stores}
-
-      Append-only stores are store where it is possible to read and add new
-      values. *)
-
-  include Read_only.S
-  (** @inline *)
-
-  val add : [> write ] t -> key -> value -> unit Lwt.t
-  (** Write the contents of a value to the store. *)
-
-  include Clearable with type 'a t := 'a t
-end
-
-module type Maker = functor (K : Type.S) (V : Type.S) -> sig
-  include S with type key = K.t and type value = V.t
-  include Batch with type 'a t := 'a t
-  include Of_config with type 'a t := 'a t
-  include Closeable with type 'a t := 'a t
-end
-
-module type Sigs = sig
-  module type S = S
-  module type Maker = Maker
-end
+include Read_only_intf.Sigs